+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.X.X]
+
+### Changed
+
+* Refactor code of TargetSelector, make it a real class.
+* Rename tree view elements.
+
+### Fixed
+
+* Fix target name bug for Rust apps.
+
 ## [0.2.1]
 
 ### Added